Expand example app for enhanced test coverage (#12)
There were a number of codepaths in the spec generator that were developed against the Katapult API but were not represented in the example app contained in this repo. So now all code paths in the generator are triggered.

We can see this, because i've added Simplecov for code coverage checking.

I think this makes sense, as it will ensure that all code-paths used by the specification generator are actually executed by the example application. Which gives us greater confidence that we are generating the spec correctly for all the various ways the apia endpoints can be defined.

This also revealed a bug in the way arrays were represented for some parts of the schema. So this PR also fixes that at the same time.

I've removed a bunch of code comment TODOs because we have issues to track these now.

I have verified these changes against the Katapult API.

closes: https://github.com/krystal/apia-openapi/issues/8