implement partial polymorphs (#17)
We need to handle when the endpoint has said a polymorph will be returned, but not all the fields will be included in the response. In that case we shouldn't just point to a $ref (as that implies all fields will be returned).

This is triggered by the Katapult API. So we do need it, specifically for `IPAddresses::InfoEndpoint`

For now I have assumed that if we are selecting specific attributes from the polymorph options, that those attributes appear in all the options and use the same data type. This is currently the case for Katapult, and just keeps it simpler.

closes: #13