Handle nullable response fields
In Apia we can add `null: true` to a field, like this:

```ruby
field :allocation, type: Objects::IPAddressAllocation, include: 'id,name', null: true
```

It means that in the response it will either be an `IPAddressAllocation` object or it might not be present at all.

In OpenAPI 3.0 we can represent this in the spec by using `nullable: true`.

closes: https://github.com/krystal/apia-openapi/issues/18