Add descriptions for LookupArgumentSets
OpenAPI 3.0 does not support parameter dependencies and mutually exclusive parameters.

https://swagger.io/docs/specification/describing-parameters/#dependencies

This is a complex topic with a lot of debate:
https://github.com/OAI/OpenAPI-Specification/issues/256

We know that our LookupArgumentSets are mutually exclusive, so the best thing is to
add a description explaining this, so that it will appear in documentation and tools like the swagger editor.

Apia already returns a 400 with a friendly error message.

closes: https://github.com/krystal/apia-openapi/issues/4