Use component schemas instead of inline schemas
The generated clients try to be smart and "re-use" schemas where possible. This makes for some really odd documentation and for go and php, some really odd types.

e.g in the load balancer response, it says certificates property is type GetDataCenters200ResponseDataCentersInnerCountry which has nothing to do with the certificates. It's just that it uses the same properties.

Now we don't define any schemas "inline" and always point to a $ref. The idea being that the generators won't try to be smart and optimise the types.

closes: https://github.com/krystal/apia-openapi/issues/22