Shorten ids for components
This makes all the classes and types far less "wordy" in the generated code.

This was especially a problem in the generated go client code, which was producing exceptionally long names for enum consts that cannot clash across the entire api.

This change does introduce a risk that two unrelated things could end up with the same id.
In the Katapult API this doesn't appear to be the case.

If an API introduced this, we could either:
1. rename the conflicting thing to something unique
2. introduce some kind of "collision detection", where we check the full id against the shortened id. and don't re-use the id if it already points to something else.