Improve error response names
Currently we declare the same 403 response many times over for many endpoints. Just with a unique name for the endpoint.
Same for other http status codes.

We did that because for other parts of the code generator, generating unique component refs forced the generator not to be "smart" and re-use things across different endpoints.

However, this is not the case with the error responses.

So DeleteVirtualMachine403Response is actually being used for ~43 different api calls. Not obviously all related to deleting a virtual machine.

Now we name the error response after the errors it returns (excluding the api authenticator ones applied to all responses).

closes: https://github.com/krystal/apia-openapi/issues/28