Improve error response names (#29)
Currently we declare the same 403 response many times over for many endpoints. Just with a unique name for the endpoint. Same for other http status codes.

We did that because for other parts of the code generator, generating unique component refs forced the generator not to be "smart" and re-use things across different endpoints.

However, this is not the case with the error responses.

So DeleteVirtualMachine403Response is actually being used for ~43 different api calls. Not obviously all related to deleting a virtual machine.

Now we name the error response after the errors it returns (excluding the api authenticator ones applied to all responses).

In addition, we don't declare the `oneOf` inside the response, that has been moved to another named schema definition and then referenced from the response. This forces the generator to use the named schema and not try to be "clever".

closes: https://github.com/krystal/apia-openapi/issues/28