feat: limit the generated names of the $refs
If we use a generator to produce a Ruby client, and then try to build that gem, it's possible for the RubyGems build process to fail with:
`ERROR: While executing gem ... (Gem::Package::TooLongFileName)`
This can happen if the $ref ids we are using are too long.

This PR makes 3 changes which are applied during the generation of ids only if they too long:
- error responses that can be "one of" many, will now no longer use all of the error names to generate the id
- nested partial responses will no longer use all of the parent field names in the id generation
- endpoint ids will resort to using the initials of the full path, when the id has a clash and the path is too long (this will be super rare).

this is a breaking change, but we're on 0.1.1 and not actually using any generated clients yet. So I'm not marking it as such.

closes: #46